chore: Refactor Player component to use PlayersContext for managing players state and implement delete player functionality

+++ app/frontend/src/components/player.tsx
@@ -1,24 +1,66 @@
+import { useContext } from 'react';
 import { PlayerType } from '../types/PlayerType';
+import PlayersContext from '../context/PlayersContext';
+import Swal from 'sweetalert2';
 
-export default function Player (player: PlayerType) {
-  const { nickname, missedWars, player_id, class: classe} = player;
+export default function Player(player: PlayerType) {
+	const { nickname, missedWars, player_id, class: classe } = player;
+
+	const playersContext = useContext(PlayersContext);
+	if (!playersContext) {
+		throw new Error('SomeComponent must be used within a PlayersProvider');
+	}
+	const { players, setPlayers } = playersContext;
 
 	const handleDelete = async () => {
-		console.log(player);
-		
-		try {
-			const response = await fetch(`http://localhost:3001/players/${player_id}`, {
-				method: 'DELETE',
-			});
-			if (response.ok) {
-				console.log('Player deletado com sucesso!');
-			} else {
-				console.error('Erro ao deletar player!');
+		const result = await Swal.fire({
+			title: 'Você tem certeza?',
+			text: "Você não poderá reverter isso!",
+			icon: 'warning',
+			showCancelButton: true,
+			confirmButtonColor: '#3085d6',
+			cancelButtonColor: '#d33',
+			confirmButtonText: 'Sim, deletar!',
+		});
+
+		if (result.isConfirmed) {
+			try {
+				const response = await fetch(
+					`http://localhost:3001/players/${player_id}`,
+					{
+						method: 'DELETE',
+					}
+				);
+				if (response.ok) {
+					console.log('Player deletado com sucesso!');
+					Swal.fire({
+						title: 'Deleted!',
+						text: 'Your file has been deleted.',
+						icon: 'success',
+					});
+					const newPlayers = players.filter(
+						(player: PlayerType) => player.player_id !== player_id
+					);
+					setPlayers(newPlayers);
+				} else {
+					Swal.fire({
+						icon: "error",
+						title: "Oops...",
+						text: "Erro ao deletar jogador!",
+					});
+					console.error('Erro ao deletar player!');
+				}
+			} catch (error) {
+				Swal.fire({
+					icon: "error",
+					title: "Oops...",
+					text: "Algo deu errado! Tente novamente!",
+				});
+				console.error('Error:', error);
 			}
-		} catch (error) {
-			console.error('Error:', error);
 		}
 	};
+
 	return (
 		<div className='player' key={player_id}>
 			<h3>{nickname}</h3>
